- Initial import of the mediaproxy nat traversal module into cvs. - Check the Readme file for a description of the architecture and   functionality provided by the module. - Also check http://mediaproxy.ag-projects.com/ from where you can   download the SER Mediaproxy server that works in conjunction with   this module. The README and INSTALL files from SER Mediaproxy   provide more indepth information about the architecture they impelemnt   and the way they are supposed to work together.   Visit http://ag-projects.com/SER_Media_Proxy.html for more details   about SER Mediaproxy.
To sumarize the features provided by mediaproxy (module+proxy server)
we can say that it is designed as a nat traversal solution (implements
the same idea of using a proxy in the middle as the nathelper module).

However unlike the nathelper module this solution will allow one to
configure local or remote proxy servers based on the SIP domain of the
caller/destination using SRV records in DNS.
Even more there can be multiple proxy servers configured for a domain,
providing load balancing of the RTP traffic between all defined proxy
servers using the weight property of the SRV records.
If one doesn't desire proxy servers to be configured per domain, but
wants all calls no matter on which domain they are to be spread among
multiple proxy servers that guarantee load balancing and fallback
features, this can be done too.
For more details read the README files from both the SER mediaproxy
module as well as from the SERMediaproxy server.