Iptrtpproxy  provides  similar functionality as nathelper but communicates with netfilter kernel xt_RTPPROXY module using libipt_RTPPROXY userspace library. See http://www.2p.cz/en/netfilter_rtp_proxy All RTP streams are manipulated directly in kernel space, no data are copied from kernel to userspace and back, it reduces load and delay.
The iptrtpproxy module is written as light-weighted, there is not implemented any
dialog managment as is e.g. in nathelper, the reason is that such API should be
provided by core or specialized dialog manager module. Because such module
is not in CVS, session information may be stored in extra attributes of
avp_db  module and session id itself in record route as cookie, see rr
module.