core: compile io_wait support even if -DUSE_TCP is not set
io_wait is needed now also by the ctl module, so it should be
compiled even if no tcp support is desired.
Its compilation can be turned off if NO_IO_WAIT is defined
(but in this case one must make sure so no other piece of code
using it will be compiled).